+++ src/main/java/com/kh/workPeople/configuration/CommonSecurityConfiguration.java
@@ -1,78 +1,16 @@
 package com.kh.workPeople.configuration;
 
-import java.io.IOException;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-import com.kh.workPeople.account.login.model.service.LoginService;
 
 @Configuration
-@EnableWebSecurity
-public class CommonSecurityConfiguration extends WebSecurityConfigurerAdapter  {
-	
-	private LoginService loginService;
-	
-	@Autowired
-	public CommonSecurityConfiguration(LoginService loginService) {
-		this.loginService = loginService;
-	}
-	
-	public static final String[] SECURITY_EXCLUDE_PATTERN = {
-		   "/css/**"
-		 , "/js/**"
-		 , "/images/**"
-	};
-	
+public class CommonSecurityConfiguration  {
+
 	@Bean
 	public PasswordEncoder passwordEncoder() {
 		return new BCryptPasswordEncoder();
 	}
 
-	@Override
-	public void configure(WebSecurity web) throws Exception {
-		web.ignoring().antMatchers(SECURITY_EXCLUDE_PATTERN);
-	}
-
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		http
-			.authorizeRequests()
-				.anyRequest().permitAll()
-			.and()
-				.formLogin()
-				.loginPage("/account/login")
-				.usernameParameter("id")
-				.passwordParameter("password")
-				.successForwardUrl("/main")
-			.and()
-				.logout()
-				.logoutRequestMatcher(new AntPathRequestMatcher("/account/logout"))
-				.deleteCookies("JESSIONID")
-				.invalidateHttpSession(true)
-				.logoutSuccessUrl("/main");
-	}
-
-
-	@Override
-	public void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.userDetailsService(loginService).passwordEncoder(passwordEncoder());
-	}
 }
